updated to avoid new compiler sensitivity to type mismatches in printf

+++ src/pd/red-black/rb.c
@@ -135,7 +135,8 @@ Rb_node rb_find_key_n(Rb_node n, char *key, int *fnd)
  
   *fnd = 0;
   if (!ishead(n)) {
-    fprintf(stderr, "rb_find_key_n called on non-head 0x%x\n", n);
+    fprintf(stderr, "rb_find_key_n called on non-head 0x%x\n", 
+	    (unsigned int) n);
     exit(1);
   }
   if (n->p.root == n) return n;
@@ -167,7 +168,8 @@ Rb_node rb_find_ikey_n(Rb_node n, unsigned int ikey, int *fnd)
 {
   *fnd = 0;
   if (!ishead(n)) {
-    fprintf(stderr, "rb_find_ikey_n called on non-head 0x%x\n", n);
+    fprintf(stderr, "rb_find_ikey_n called on non-head 0x%x\n", 
+	    (unsigned int) n);
     exit(1);
   }
   if (n->p.root == n) return n;
@@ -199,7 +201,8 @@ Rb_node rb_find_gkey_n(Rb_node n, char *key,int (*fxn)(char *, char *), int *fnd
  
   *fnd = 0;
   if (!ishead(n)) {
-    fprintf(stderr, "rb_find_key_n called on non-head 0x%x\n", n);
+    fprintf(stderr, "rb_find_key_n called on non-head 0x%x\n", 
+	    (unsigned int) n);
     exit(1);
   }
   if (n->p.root == n) return n;
@@ -356,11 +359,13 @@ void rb_delete_node(Rb_node n)
   char ir;
  
   if (isint(n)) {
-    fprintf(stderr, "Cannot delete an internal node: 0x%x\n", n);
+    fprintf(stderr, "Cannot delete an internal node: 0x%x\n", 
+	    (unsigned int) n);
     exit(1);
   }
   if (ishead(n)) {
-    fprintf(stderr, "Cannot delete the head of an rb_tree: 0x%x\n", n);
+    fprintf(stderr, "Cannot delete the head of an rb_tree: 0x%x\n", 
+	    (unsigned int) n);
     exit(1);
   }
   delete_item(n); /* Delete it from the list */
@@ -476,23 +481,25 @@ void rb_print_tree(Rb_node t, int level)
 {
   int i;
   if (ishead(t) && t->p.parent == t) {
-    printf("tree 0x%x is empty\n", t);
+    printf("tree 0x%x is empty\n", (unsigned int) t);
   } else if (ishead(t)) {
-    printf("Head: 0x%x.  Root = 0x%x\n", t, t->p.root);
+    printf("Head: 0x%x.  Root = 0x%x\n", (unsigned int) t, 
+	   (unsigned int) t->p.root);
     rb_print_tree(t->p.root, 0);
   } else {
     if (isext(t)) {
       for (i = 0; i < level; i++) putchar(' ');
       printf("Ext node 0x%x: %c,%c: p=0x%x, k=%s\n", 
-              t, isred(t)?'R':'B', isleft(t)?'l':'r', t->p.parent, t->k.key);
+	     (unsigned int) t, isred(t)?'R':'B', isleft(t)?'l':'r', 
+	     (unsigned int) t->p.parent, t->k.key);
     } else {
       rb_print_tree(t->c.child.left, level+2);
       rb_print_tree(t->c.child.right, level+2);
       for (i = 0; i < level; i++) putchar(' ');
       printf("Int node 0x%x: %c,%c: l=0x%x, r=0x%x, p=0x%x, lr=(%s,%s)\n", 
-              t, isred(t)?'R':'B', isleft(t)?'l':'r', t->c.child.left, 
-              t->c.child.right, 
-              t->p.parent, t->k.lext->k.key, t->v.rext->k.key);
+	     (unsigned int) t, isred(t)?'R':'B', isleft(t)?'l':'r', 
+	     (unsigned int) t->c.child.left, (unsigned int) t->c.child.right, 
+             (unsigned int) t->p.parent, t->k.lext->k.key, t->v.rext->k.key);
     }
   }
 }
@@ -501,25 +508,27 @@ void rb_iprint_tree(Rb_node t, int level)
 {
   int i;
   if (ishead(t) && t->p.parent == t) {
-    printf("tree 0x%x is empty\n", t);
+    printf("tree 0x%x is empty\n", (unsigned int) t);
   } else if (ishead(t)) {
     printf("Head: 0x%x.  Root = 0x%x, < = 0x%x, > = 0x%x\n", 
-            t, t->p.root, t->c.list.blink, t->c.list.flink);
+	   (unsigned int) t, (unsigned int) t->p.root, 
+	   (unsigned int) t->c.list.blink, (unsigned int) t->c.list.flink);
     rb_iprint_tree(t->p.root, 0);
   } else {
     if (isext(t)) {
       for (i = 0; i < level; i++) putchar(' ');
       printf("Ext node 0x%x: %c,%c: p=0x%x, <=0x%x, >=0x%x k=%d\n", 
-              t, isred(t)?'R':'B', isleft(t)?'l':'r', t->p.parent, 
-              t->c.list.blink, t->c.list.flink, t->k.ikey);
+	     (unsigned int) t, isred(t)?'R':'B', isleft(t)?'l':'r', 
+	     (unsigned int) t->p.parent, (unsigned int) t->c.list.blink, 
+	     (unsigned int) t->c.list.flink, t->k.ikey);
     } else {
       rb_iprint_tree(t->c.child.left, level+2);
       rb_iprint_tree(t->c.child.right, level+2);
       for (i = 0; i < level; i++) putchar(' ');
       printf("Int node 0x%x: %c,%c: l=0x%x, r=0x%x, p=0x%x, lr=(%d,%d)\n", 
-              t, isred(t)?'R':'B', isleft(t)?'l':'r', t->c.child.left, 
-              t->c.child.right, 
-              t->p.parent, t->k.lext->k.ikey, t->v.rext->k.ikey);
+	     (unsigned int) t, isred(t)?'R':'B', isleft(t)?'l':'r', 
+	     (unsigned int) t->c.child.left, (unsigned int) t->c.child.right, 
+             (unsigned int) t->p.parent, t->k.lext->k.ikey, t->v.rext->k.ikey);
     }
   }
 }
@@ -529,7 +538,7 @@ int rb_nblack(Rb_node n)
   int nb;
   if (ishead(n) || isint(n)) {
     fprintf(stderr, "ERROR: rb_nblack called on a non-external node 0x%x\n",
-            n);
+            (unsigned int) n);
     exit(1);
   }
   nb = 0;
@@ -545,7 +554,7 @@ int rb_plength(Rb_node n)
   int pl;
   if (ishead(n) || isint(n)) {
     fprintf(stderr, "ERROR: rb_plength called on a non-external node 0x%x\n",
-            n);
+            (unsigned int) n);
     exit(1);
   }
   pl = 0;
